requirements.txt
Django==1.5.3
South==0.8.2
argparse==1.2.1
dj-database-url==0.2.2
dj-static==0.0.5
django-allauth==0.13.0
django-pylibmc-sasl==0.2.4
django-registration==1.0
django-toolbelt==0.0.1
gunicorn==18.0
oauthlib==0.6.0
psycopg2==2.5.1
pylibmc==1.2.3
python-dateutil==2.1
pytz==2013.7
six==1.4.1
static==0.4
wsgiref==0.1.2
